+++ tools/bin/main.ml
@@ -7,14 +7,23 @@ Usage: rescript-tools doc <FILE>
 
 Example: rescript-tools doc ./path/to/EntryPointLib.res|}
 
-let formatDocstringsHelp =
+let formatCodeblocksHelp =
   {|ReScript Tools
 
-Format ReScript code blocks in docstrings
+Format ReScript code blocks in docstrings or markdown files
 
-Usage: rescript-tools format-docstrings <FILE> [--stdout] [--transform-assert-equal]
+Usage: rescript-tools format-codeblocks <FILE> [--stdout] [--transform-assert-equal]
 
-Example: rescript-tools format-docstrings ./path/to/MyModule.res|}
+Example: rescript-tools format-codeblocks ./path/to/MyModule.res|}
+
+let extractCodeblocksHelp =
+  {|ReScript Tools
+
+Extract ReScript code blocks from docstrings or markdown files
+
+Usage: rescript-tools extract-codeblocks <FILE> [--transform-assert-equal]
+
+Example: rescript-tools extract-codeblocks ./path/to/MyModule.res|}
 
 let help =
   {|ReScript Tools
@@ -27,6 +36,8 @@ doc <file>                              Generate documentation
 format-codeblocks <file>                Format ReScript code blocks
   [--stdout]                              Output to stdout
   [--transform-assert-equal]              Transform `assertEqual` to `==`
+extract-codeblocks <file>               Extract ReScript code blocks from file
+  [--transform-assert-equal]              Transform `==` to `assertEqual`
 reanalyze                               Reanalyze
 -v, --version                           Print version
 -h, --help                              Print help|}
@@ -57,7 +68,7 @@ let main () =
     | _ -> logAndExit (Error docHelp))
   | "format-codeblocks" :: rest -> (
     match rest with
-    | ["-h"] | ["--help"] -> logAndExit (Ok formatDocstringsHelp)
+    | ["-h"] | ["--help"] -> logAndExit (Ok formatCodeblocksHelp)
     | path :: args -> (
       let isStdout = List.mem "--stdout" args in
       let transformAssertEqual = List.mem "--transform-assert-equal" args in
@@ -71,7 +82,26 @@ let main () =
       | Ok content, `Stdout -> print_endline content
       | result, `File -> logAndExit result
       | Error e, _ -> logAndExit (Error e))
-    | _ -> logAndExit (Error formatDocstringsHelp))
+    | _ -> logAndExit (Error formatCodeblocksHelp))
+  | "extract-codeblocks" :: rest -> (
+    match rest with
+    | ["-h"] | ["--help"] -> logAndExit (Ok extractCodeblocksHelp)
+    | path :: args -> (
+      let transformAssertEqual = List.mem "--transform-assert-equal" args in
+      Clflags.color := Some Misc.Color.Never;
+
+      (* TODO: Add result/JSON mode *)
+      match
+        Tools.ExtractCodeblocks.extractCodeblocksFromFile ~transformAssertEqual
+          ~entryPointFile:path
+      with
+      | Ok _ as r ->
+        print_endline (Analysis.Protocol.stringifyResult r);
+        exit 0
+      | Error _ as r ->
+        print_endline (Analysis.Protocol.stringifyResult r);
+        exit 1)
+    | _ -> logAndExit (Error extractCodeblocksHelp))
   | "reanalyze" :: _ ->
     let len = Array.length Sys.argv in
     for i = 1 to len - 2 do